add subComponent
