+++ src/components/article-picker/index.js
@@ -1,0 +1,28 @@
+import React, { useMemo } from 'react';
+import SearchArticlesSort from '@triniti/schemas/triniti/news/enums/SearchArticlesSort.js';
+import withRequest from '@triniti/cms/plugins/pbjx/components/with-request/index.js';
+import NodePickerField from '@triniti/cms/plugins/ncr/components/node-picker-field/index.js';
+
+const ArticlePicker = (props) => {
+  const { statuses, ...rest } = props;
+
+  const ArticlePickerField = useMemo(() => {
+    const initialData = {
+      sort: SearchArticlesSort.ORDER_DATE_DESC.getValue(),
+      autocomplete: true
+    };
+
+    if (statuses) {
+      initialData.statuses = statuses;
+    }
+
+    return withRequest(NodePickerField, 'triniti:news:request:search-articles-request', {
+      channel: 'picker',
+      initialData
+    });
+  }, [statuses?.length, statuses?.join(',')]);
+
+  return <ArticlePickerField {...rest} />;
+};
+
+export default ArticlePicker;
