+++ src/plugins/ovp/components/video-picker-field/index.js
@@ -1,28 +1,6 @@
-import React, { useMemo } from 'react';
-import SearchVideosSort from '@triniti/schemas/triniti/ovp/enums/SearchVideosSort.js';
-import withRequest from '@triniti/cms/plugins/pbjx/components/with-request/index.js';
-import NodePickerField from '@triniti/cms/plugins/ncr/components/node-picker-field/index.js';
+import React from 'react';
+import VideoPicker from '@triniti/cms/components/video-picker/index.js';
 
-const VideoPickerFieldWithStatus = (props) => {
-  const { statuses, ...rest } = props;
-
-  const VideoPickerField = useMemo(() => {
-    const initialData = {
-      sort: SearchVideosSort.ORDER_DATE_DESC.getValue(),
-      autocomplete: true
-    };
-
-    if (statuses) {
-      initialData.statuses = statuses;
-    }
-
-    return withRequest(NodePickerField, 'triniti:ovp:request:search-videos-request', {
-      channel: 'picker',
-      initialData
-    });
-  }, [statuses?.length, statuses?.join(',')]);
-
-  return <VideoPickerField {...rest} />;
+export default function VideoPickerField(props) {
+  return <VideoPicker {...props} />;
 };
-
-export default VideoPickerFieldWithStatus;
